Refactor imports and update version metadata
